Implemented dynamic car filtering, sorting, multi-image slider, and YouTube trailer player.
DEPENDENCIES (build.gradle):
  *  Added Coil (`coil-compose`) for network image loading.

  *  Added `androidyoutubeplayer:core` for embedded video playback.

DATA LAYER (Room Database & DTOs):
  *  Created a `Converters` class to store lists of image URLs as comma-separated text inside Room.

  *  Added `@TypeConverters` to `AppDatabase` and updated `CarEntity` to support `List<String>` for image fields.

  *  Changed data types for better accuracy: `engine` from String to Int, and `consumption` from String to Double across database entities, domain models, and network DTOs.

  *  Updated ratings to Float (`ReviewEntity`, `ReviewRequest`, `VehicleReview`) to support half-star precise ratings.

  *  Implemented local database cleanup (`deleteAll()`) in `RestCarRepository` when active filters are used, preventing outdated cached data from showing up.

REMOTE API & SERVICE LAYER:
  *  Expanded `CarEntryApiService` and `CarRepository.refreshCars()` to support 12 optional query parameters for filtering data directly from the Django backend (price, CC, mileage, HP, year, category, and sorting).

  *  Added remote deletion support via `deleteCarEntry` and handled cascade local deletion for vehicles and their reviews.

UI COMPONENTS & CUSTOM VIEWS:
  *  Created a reusable `YoutubePlayer` component using Android's native view wrapper, safely tied to the Activity/Fragment lifecycle.

  *  Added `extractYoutubeVideoId` helper function to parse and extract the 11-character ID from both desktop and mobile shortened YouTube links.

  *  Updated `StarRatingBar` to allow users to click and change the rating when a callback function is provided.

  *  Created `CarListUiState` and `CarFilterState` to manage loading states, error messages, and active filter values inside `CarListViewModel`.

MAINACTIVITY, SCREENS & NAVIGATION:
  *  Integrated a `ModalBottomSheet` inside `VehiclesScreen` to house all the filter input fields (Min/Max limits).

  *  Added a dropdown menu overlay for sorting options, mapping user-friendly labels to Django backend keys (`price`, `-price`, `-year`, `engine`).

  *  Implemented a dynamic image slider inside `VehicleDetailScreen` using `HorizontalPager` and Coil `AsyncImage`, complete with a text indicator overlay (e.g., "1 / 3").

  * Embedded the custom YouTube player inside the vehicle details layout to preview car trailers.

  *  Used `LaunchedEffect` inside `CarShowNavHost` to automatically refresh and apply filters whenever the navigation route category changes.

+++ app/src/main/java/gr/ihu/ict/carshow/ui/components/ReviewComponents.kt
@@ -40,9 +40,9 @@ import androidx.compose.ui.unit.sp
 // If not pass it the bar is read-only (just for showing the score)
 @Composable
 fun StarRatingBar(
-    rating: Double,
+    rating: Float,
     maxStars: Int = 5,
-    onRatingChanged: ((Double) -> Unit)? = null // Nullable callback determines if the bar is interactive
+    onRatingChanged: ((Float) -> Unit)? = null // Nullable callback determines if the bar is interactive
 ) {
     Row(verticalAlignment = Alignment.CenterVertically) {
         for (i in 1..maxStars) {
@@ -62,7 +62,7 @@ fun StarRatingBar(
                     .then(
                         // If a callback is provided, make the star clickable to update rating (only if want user to change the rating)
                         if (onRatingChanged != null) {
-                            Modifier.clickable{ onRatingChanged(i.toDouble()) }
+                            Modifier.clickable{ onRatingChanged(i.toFloat()) }
                         } else {
                             Modifier
                         }
@@ -86,7 +86,7 @@ fun StarRatingBar(
 @Composable
 fun ReviewItem(
     username: String,
-    rating: Double,
+    rating: Float,
     comment: String,
     date: String
 ) {
@@ -145,10 +145,10 @@ fun ReviewItem(
 @Composable
 fun AddReviewDialog(
     onDismiss: () -> Unit,
-    onConfirm: (Double, String) -> Unit
+    onConfirm: (Float, String) -> Unit
 ) {
     // Temporary Local state to hold user input (what user selects/types) before hitting submit button
-    var rating by remember{ mutableStateOf(0.0) }
+    var rating by remember{ mutableStateOf(0.0f) }
     var comment by remember { mutableStateOf("") }
 
 
@@ -194,7 +194,7 @@ fun AddReviewDialog(
                 // Trigger callback for sending the selected rating and text comment back when clicked
                 onClick = { onConfirm(rating, comment) },
                 // Validation, disable button if no rating is given or comment is empty
-                enabled = rating > 0 && comment.isNotBlank()
+                enabled = rating > 0f && comment.isNotBlank()
             ) {
                 Text("Submit")
             }
